Break out the code to linkify the byline into a common component
