+++ src/web/components/Contributor.tsx
@@ -1,9 +1,13 @@
 import React from 'react';
 import { css } from 'emotion';
-import TwitterIcon from '@frontend/static/icons/twitter.svg';
+
 import { palette } from '@guardian/src-foundations';
 import { headline, textSans } from '@guardian/src-foundations/typography';
 import { pillarPalette } from '@root/src/lib/pillars';
+
+import { BylineLink } from '@frontend/web/components/BylineLink';
+
+import TwitterIcon from '@frontend/static/icons/twitter.svg';
 
 const twitterHandle = css`
     ${textSans.xsmall()};
@@ -43,55 +47,6 @@ const bylineStyle = (pillar: Pillar) => css`
     }
 `;
 
-// This crazy function aims to split bylines such as
-// 'Harry Potter in Hogwarts' to ['Harry Potter', 'in Hogwarts']
-// Or
-// 'Jane Doe and John Smith` to ['Jane Doe', ' and ', 'John Smith']
-// It does this so we can have separate links to both contributors
-const bylineAsTokens = (bylineText: string, tags: TagType[]): string[] => {
-    const contributorTags = tags
-        .filter(t => t.type === 'Contributor')
-        .map(c => c.title);
-    // The contributor tag title should exist inside the bylineText for this regex to work
-    const regex = new RegExp(`(${contributorTags.join('|')})`);
-
-    return bylineText.split(regex);
-};
-
-const RenderContributor: React.FC<{
-    bylineText: string;
-    tags: TagType[];
-}> = ({ bylineText, tags }) => {
-    const renderedTokens = bylineAsTokens(bylineText, tags).map((token, i) => {
-        const associatedTags = tags.filter(t => t.title === token);
-        if (associatedTags.length > 0) {
-            return (
-                <ContributorLink
-                    contributor={token}
-                    contributorTagId={associatedTags[0].id}
-                    key={i}
-                />
-            );
-        }
-        return token;
-    });
-
-    return <>{renderedTokens}</>;
-};
-
-const ContributorLink: React.FC<{
-    contributor: string;
-    contributorTagId: string;
-}> = ({ contributor, contributorTagId }) => (
-    <a
-        rel="author"
-        data-link-name="auto tag link"
-        href={`//www.theguardian.com/${contributorTagId}`}
-    >
-        {contributor}
-    </a>
-);
-
 export const Contributor: React.FC<{
     author: AuthorType;
     tags: TagType[];
@@ -104,7 +59,7 @@ export const Contributor: React.FC<{
     return (
         <address aria-label="Contributor info">
             <div className={bylineStyle(pillar)}>
-                <RenderContributor bylineText={author.byline} tags={tags} />
+                <BylineLink byline={author.byline} tags={tags} />
             </div>
             {author.twitterHandle && (
                 <div className={twitterHandle}>
